% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_tidy.R
\name{fetch_and_tidy}
\alias{fetch_and_tidy}
\title{Fetch and Tidy Data}
\usage{
fetch_and_tidy(tbl_name, connection)
}
\arguments{
\item{tbl_name}{Name of data table}

\item{connection}{Database connection object}
}
\value{
A tibble of tidy data
}
\description{
Fetches Access tables and performs some basic cleaning of white space and NA strings
}
