% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_access.R
\name{fetch_from_access}
\alias{fetch_from_access}
\title{Fetch data from Access database}
\usage{
fetch_from_access(
  db_path,
  data_prefix = "qExport",
  lookup_prefix = "tlu",
  tables_to_omit = c(),
  custom_wrangler,
  save_to_files = FALSE,
  data_dir = here::here("data", "final"),
  dictionary_dir = here::here("data", "dictionary")
)
}
\arguments{
\item{db_path}{Path to your Access database}

\item{data_prefix}{Prefix used in your Access database to indicate data export tables and/or queries}

\item{lookup_prefix}{Prefix used in your Access database to indicate lookup tables}

\item{tables_to_omit}{Character vector of table names that match the data}

\item{custom_wrangler}{Optional - function that takes arguments `data`, `lookups`, and `metadata`. All are lists whose names and content correspond to the export data, lookup, and metadata tables in the database after using the database tool.}

\item{save_to_files}{Should the function save data and data dictionaries to files on hard drive?}

\item{data_dir}{Folder to store data csv's in}

\item{dictionary_dir}{Folder to store data dictionaries in}
}
\value{
A nested list containing three lists of tibbles: data, lookups, and metadata.
}
\description{
This function assumes that you are using the Access Metadata Generator in your Access database.It is the main function to get the tables from the user defined access database, perform data wrangling, and produce the dicts needed to create eml metadata using the NPS eml tools as part of a data package. This function should perform any necessary data wrangling specific to your dataset and return a named list containing `data`, `lookups`, and `metadata` with contents modified as needed. Do not remove or add tibbles in `data` or `lookups` and do not modify their names. If you add, remove, or rename columns in a tibble in `data`, you must modify the contents of `metadata` accordingly. Do not modify the structure or column names of `metadata`. The structure and column names of `lookups` should also be left as-is. Typically the only necessary modification to `lookups` will be to filter overly large species lists to only include taxa that appear in the data.
}
