% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchaccess-package.R
\docType{package}
\name{fetchaccess-package}
\alias{fetchaccess}
\alias{fetchaccess-package}
\title{fetchaccess: Fetch Data And Metadata From Access}
\description{
Reads data and metadata from an Access database using the metadata generator tool.
}
\author{
\strong{Maintainer}: Sarah Wright \email{sarah_wright@nps.gov}

}
\keyword{internal}
