% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{fetchAndTidy}
\alias{fetchAndTidy}
\title{Fetch and tidy data}
\usage{
fetchAndTidy(tbl_name, connection, as.is)
}
\arguments{
\item{tbl_name}{Name of data table}

\item{connection}{Database connection object}
}
\value{
A tibble of tidy data
}
\description{
Fetch and tidy data
}
