% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_tables.R
\name{get_access_tables}
\alias{get_access_tables}
\title{Get Access Tables}
\usage{
get_access_tables(db_path, data_prefix, lookup_prefix, tables_to_omit)
}
\arguments{
\item{db_path}{Path to your Access database}

\item{data_prefix}{Prefix used in your Access database to indicate data export tables and/or queries}

\item{lookup_prefix}{Prefix used in your Access database to indicate lookup tables}

\item{tables_to_omit}{a list of any tables to omit from the import}
}
\value{
A nested list containing three lists of tibbles: data, lookups, and metadata.
}
\description{
This function assumes that you are using the Access Metadata Generator in your Access database, and using the fetch_from_access function. It also assumes that the user has 64bit Access, 64 bit R installed, and the database version is newer than 2007 and not password protected. It will not work with 32 bit MS Access or R.  This function connects to the user designated access database and extracts the export data, lookup, and metadata tables needed.
}
