% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{getRClass}
\alias{getRClass}
\title{Helper function to get primary R class of each data column}
\usage{
getRClass(fields, data)
}
\arguments{
\item{fields}{Fields data dictionary, as returned by \link{fetchFromAccess}}

\item{data}{List of data tables, as returned by \link{fetchFromAccess}}
}
\value{
\code{fields} with an additional rClass column
}
\description{
Helper function to get primary R class of each data column
}
