% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_template.R
\name{getTemplateDatabase}
\alias{getTemplateDatabase}
\title{Get the metadata generator template database}
\usage{
getTemplateDatabase(destination_folder, overwrite_existing = FALSE)
}
\arguments{
\item{destination_folder}{Where to save the template database}

\item{overwrite_existing}{If the template database already exists in the destination folder, should it be overwritten?}
}
\description{
Copies the metadata generator template database to a location of your choice.
}
