% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{makeColSpec}
\alias{makeColSpec}
\title{Generate column spec from data dictionary}
\usage{
makeColSpec(fields)
}
\arguments{
\item{fields}{Fields data dictionary, as returned by [fetchFromAccess]}
}
\value{
A list of lists
}
\description{
Given a fields data dictionary, create a list of column specifications that can be used in [readr::read_csv()] or [vroom::vroom()]
}
