% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_metadata.R
\name{test_metadata}
\alias{test_metadata}
\title{Test Metadata}
\usage{
test_metadata(metadata)
}
\arguments{
\item{metadata}{list of metadata tables created as part of the fetch_from_access function}
}
\value{
a numeric value based on the number of tests the input data fails
}
\description{
function to test that thr metadata tables created in Access as part of the tool, have no missing values. Missing values at this stage would mean that the EML processes would fail later on
}
