% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_files.R
\name{write_to_files}
\alias{write_to_files}
\title{Write To Files}
\usage{
write_to_files(
  all_tables,
  data_dir = here::here("data", "final"),
  dictionary_dir = here::here("data", "dictionary")
)
}
\arguments{
\item{all_tables}{Output of `fetchFromAccess()` and `create_data_dictionaries()`}

\item{data_dir}{Folder to store data csv's in}

\item{dictionary_dir}{Folder to store data dictionaries in}
}
\description{
Write data and data dictionaries to files
}
