% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_dictionaries.R
\name{create_data_dictionaries}
\alias{create_data_dictionaries}
\title{Create Data Dictionaries}
\usage{
create_data_dictionaries(data, lookups, metadata)
}
\arguments{
\item{data}{list containing data tables}

\item{lookups}{list containing lookup tables}

\item{metadata}{list containing metdata tables}
}
\value{
A nested list containing three lists of tibbles: data, lookups, and metadata.
}
\description{
This function assumes that you are using the Access Metadata Generator in your Access database and using the fetch_from_access function.  This function takes the lists of data, lookup, and metadata tables and performs data wrangling to get into right format for eml
}
