% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_tables.R
\name{get_access_tables}
\alias{get_access_tables}
\title{Get Access Tables}
\usage{
get_access_tables(db_path, data_prefix, lookup_prefix)
}
\arguments{
\item{db_path}{Path to your Access database}

\item{data_prefix}{Prefix used in your Access database to indicate data export tables and/or queries}

\item{lookup_prefix}{Prefix used in your Access database to indicate lookup tables}
}
\value{
A nested list containing three lists of tibbles: data, lookups, and metadata.
}
\description{
This function assumes that you are using the Access Metadata Generator in your Access database, and using the fetch_from_access function. This function connects to the user designated access database and extracts the tables needed.
}
