% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{writeToFiles}
\alias{writeToFiles}
\title{Write data and data dictionaries to files}
\usage{
writeToFiles(
  all_tables,
  data_dir = here::here("data", "final"),
  dictionary_dir = here::here("data", "dictionary")
)
}
\arguments{
\item{all_tables}{Output of `fetchFromAccess()`}

\item{data_dir}{Folder to store data csv's in}

\item{dictionary_dir}{Folder to store data dictionaries in}
}
\description{
Write data and data dictionaries to files
}
